net-misc/icaclient: Fixed src_install() for x86
Closes: https://bugs.gentoo.org/747331
Package-Manager: Portage-3.0.9, Repoman-3.0.2

+++ net-misc/icaclient/icaclient-20.06.0.15.ebuild
@@ -98,7 +98,9 @@ src_install() {
 	doexe *.DLL libproxy.so wfica AuthManagerDaemon PrimaryAuthManager selfservice ServiceRecord
 
 	exeinto "${ICAROOT}"/lib
-	rm lib/ctxjpeg_fb_8.so || die
+	if use amd64 ; then
+		rm lib/ctxjpeg_fb_8.so || die
+	fi
 	doexe lib/*.so
 
 	for dest in "${ICAROOT}"{,/nls/en{,.UTF-8}} ; do
